requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
